priest_regen_mana

+++ Assets/Scripts/Players/HealthPlayer.cs
@@ -157,7 +157,7 @@ public class HealthPlayer : MonoBehaviour
         {
             Health = MaxHealth;
         }
-            ShowDamagePrefab(healthInfo.ModifiedHeal, new Color(0, 0.8f, 0, 1), new Color(0, 0.8f, 0, 0.5f));
+            ShowDamagePrefab((int)healthInfo.ModifiedHeal, new Color(0, 0.8f, 0, 1), new Color(0, 0.8f, 0, 0.5f));
             UpdateHealthBar();
             UpdateHealthBarText();
     }
